Moved functions into a LEAPMacrolib.jl
Module is LMlib.jl. Also renamed files to be parallel.
